pyproject.toml
[project]
name = "erado"
description = "Simulation suite for error mitigation via dimon erasure models."
authors = [
    { name = "Sam Griffiths", email = "sgriffiths@oqc.tech" },
]
readme = "docs/README.md"
license = "BSD-3-Clause"
license-files = ["LICENSE"]
urls = { source = "https://github.com/oqc-community/erado" }
dynamic = ["version"]
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.3",
    "scipy>=1.16.2",
    "matplotlib>=3.10.6",
    "qiskit>=2.2.1",
    "qiskit-aer-gpu-cu11>=0.17.2; sys_platform == 'linux'",
    "qiskit-aer>=0.17.2; sys_platform != 'linux'",
    "pydantic>=2.12.0",
]

[dependency-groups]
dev = [
    "poethepoet>=0.37.0",
    "pyright[nodejs]>=1.1.406",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "ruff>=0.14.0",
]
docs = [
    "furo>=2025.12.19",
    "gitpython>=3.1.46",
    "myst-parser>=5.0.0",
    "sphinx<9.0.0",  # sphinx-multiversion is currently broken for Sphinx 9
    "sphinx-autoapi>=3.7.0",
    "sphinx-multiversion>=0.2.4",
]


[build-system]
requires = ["hatchling", "uv-dynamic-versioning"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
enable = true
style = "pep440"


[tool.poe.tasks]
types = "pyright ."
pre-commit = "pre-commit run --all-files"
unit-tests = "pytest tests/"
lint = "ruff check ."
stubs = "python src/make_stubs.py qiskit._accelerate.circuit ./typings"
docs = [
    { cmd = "rm -rf docs/_build" },
    { cmd = "uv run docs/version.py" },
    { cmd = "sphinx-multiversion docs docs/_build", envfile = "version.env" },
]
docs-local = { shell = "rm -rf docs/_build && sphinx-build docs docs/_build/local" }

[tool.poe.tasks.checks]
sequence = [
    "lint",
    "types",
    "unit-tests",
]
ignore_fail = true


[tool.pyright]
typeCheckingMode = "standard"
reportImplicitOverride = "error"
exclude = [
    "typings",

    # Default pyright exclude entries:
    "**/node_modules",
    "**/__pycache__",
    "**/.*",
    ".venv",
]


[tool.ruff]
line-length = 120
extend-exclude = [
    "typings",
    "examples",
    "src/make_stubs.py",
    "docs",
]

[tool.ruff.lint]
ignore = [
    "ALL",     # Strictly opt-in to ruff functionality below
    "ANN204",  # Don't require return types on 'special' methods
]
select = [
    "E",       # Enable pycodestyle rules (line length etc.)
    "D",       # Enable pydocstyle rules
    "D401",    # Docstrings must have imperative mood
    "ANN",     # Type annotations are mandatory
    "F",       # Enable Pyflakes rules (common code errors e.g. unused imports)
    "COM",     # Standardise trailing commas
]

[tool.ruff.lint.per-file-ignores]
"test_*.py" = [
    "ANN201",  # Don't require "-> None" on every test function
]
"__init__.py" = [
    "F403",    # Allow wildcard imports in __init__.py (__all__ should be defined in such imported modules)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-annotations]
suppress-dummy-args = true
